Make package_list.html cacheable
Turn on 'XBitHack full' for packages/.  This allows apache to report a
Last-Modified: time for files using SSI, if the g+x permission bit is
turned on.

Note that the time reported is the mtime of the *including* file, not
the *included* file, so when the included content is generated (as
here), we need to also arrange to update that mtime when the included
file is updated.

Also review and remove unneeded execute permissions