Note last setup version known to work on XP
This is a bit mushy: x86_64 versions probably don't work due to
toolchain updates, but I can't test that. x86 only works from a shell
for some unknown reasons (manifest changes?). And the signature can't
be checked on XP, as it uses an unsupported crypto algorithm, so let's
draw a line here.

Also: make mention of gpg signatures explicit that they are gpg.